refactor: Const를 data layer로 이동
