+++ app/src/main/java/nextstep/github/data/Const.kt
@@ -1,4 +1,4 @@
-package nextstep.github
+package nextstep.github.data
 
 object Const {
     const val ORGANIZATION_NAME = "next-step"
